# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ClaimFlow API
  version: "1.0.0"
  description: |
    ClaimFlow — SHA claims documentation audit platform. This document is the
    **source-of-truth contract** for the public `v1` HTTP surface. A CI drift
    check (`openapi-drift.integration.test.ts`) asserts that every route the
    server registers appears here (and vice versa) and validates representative
    response payloads against these schemas, so code↔spec divergence fails the build.

    ## Response envelopes & error format
    - Success responses use the `{ data, meta? }` envelope (or a raw token/binary
      body where noted).
    - **Error format is keyed off the caller, so an integrator gets ONE shape on
      every endpoint:** machine credentials (API key or OAuth client) and the
      public machine endpoints (`/v1/claims/score`, `/v1/oauth/token`) receive
      RFC 7807 `application/problem+json`; human JWT sessions receive the
      `{ errors, meta }` envelope. The `Problem` body is a strict **superset** of
      the envelope (it also carries `errors[]` and `meta.requestId`), so a client
      written against either shape works. Error responses below are documented as
      `Problem`; envelope clients read the same `errors[]`/`meta`.

    ## Integrity
    The scoring response exposes **risk level, recommended action, public flags
    (reason code + category + severity + short message), and counts only**. It
...
servers:
  - url: /
    description: Same-origin (ports 8080 API)
...
security:
  - bearerJwt: []
  - apiKey: []
  - oauth2ClientCredentials: []
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Password login (step 1; may require MFA)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login result (tokens, or an MFA challenge)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeLoginResult" }
        "401": { $ref: "#/components/responses/EnvelopeError" }
  /v1/auth/mfa/verify:
    post:
      tags: [Auth]
      operationId: mfaVerify
      summary: Complete login by verifying a TOTP code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MfaVerifyRequest" }
      responses:
        "200":
          description: Session tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeSessionTokens" }
        "401": { $ref: "#/components/responses/EnvelopeError" }
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Rotate the refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New session tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeSessionTokens" }
        "401": { $ref: "#/components/responses/EnvelopeError" }
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the current refresh-token family
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { description: Logged out, content: { application/json: { schema: { $ref: "#/components/schemas/EnvelopeObject" } } } }
  /v1/auth/me:
    get:
      tags: [Auth]
      operationId: currentUser
      summary: Current authenticated user
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeAuthenticatedUser" }
        "401": { $ref: "#/components/responses/EnvelopeError" }
...
  /v1/auth/password/change:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Change password (enforces history); revokes sessions
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200": { description: Changed, content: { application/json: { schema: { $ref: "#/components/schemas/EnvelopeObject" } } } }
        "400": { $ref: "#/components/responses/EnvelopeError" }

  # ---------------------------------------------------------------- Claims
...
  /v1/oauth/token:
    post:
      tags: [OAuth2]
      operationId: oauthToken
      summary: OAuth2 client-credentials token (RFC 6749 §4.4)
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/OAuthTokenRequest" }
          application/json:
            schema: { $ref: "#/components/schemas/OAuthTokenRequest" }
      responses:
        "200":
          description: Access token
          headers:
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthTokenResponse" }
        "401":
          description: Invalid client (problem+json)
          content:
            application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } }
        "403":
          description: Scope not granted (problem+json)
          content:
            application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } }
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human session access token from the login/MFA flow.
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >-
        Tenant-scoped machine key (`cf_…`). May also be sent as
        `Authorization: Bearer cf_…`. Authorized by explicit scopes.
    oauth2ClientCredentials:
      type: oauth2
      description: OAuth2 client-credentials; the issued bearer token is authorized by scope.
      flows:
        clientCredentials:
          tokenUrl: /v1/oauth/token
          scopes:
            "claim:create": Create claims / score
            "audit:trigger": Trigger audits
            "export:evidence": Export evidence packs
            "dashboard:view": Read dashboards
            "case:view": Read cases
            "case:manage": Manage cases
            "system:settings": Manage keys/clients/webhooks
    metricsToken:
      type: http
      scheme: bearer
      description: Optional token gating /metrics (METRICS_AUTH_TOKEN).
...
  responses:
    EnvelopeError:
      description: >-
        Error. Machine credentials (API key / OAuth) receive
        `application/problem+json`; human JWT sessions receive the
        `{ errors, meta }` envelope. The problem body is a superset (also carries
        `errors[]` + `meta.requestId`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EnvelopeErrorBody" }
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Meta:
      type: object
      properties:
        requestId: { type: string }
        cursor: { type: [string, "null"] }
        hasMore: { type: boolean }
        total: { type: integer }
      additionalProperties: true
    ApiErrorDetail:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        field: { type: [string, "null"] }
        detail: { type: object, additionalProperties: true }
    EnvelopeErrorBody:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items: { $ref: "#/components/schemas/ApiErrorDetail" }
        meta: { $ref: "#/components/schemas/Meta" }
    EnvelopeObject:
      type: object
      required: [data]
      properties:
        data: { type: object, additionalProperties: true }
        meta: { $ref: "#/components/schemas/Meta" }
...
    Problem:
      type: object
      description: >-
        RFC 7807 problem+json. Returned to machine credentials on any endpoint and
        on the public machine endpoints. Strict superset of the envelope: also
        carries meta.requestId and errors[].
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        instance: { type: string }
        meta: { $ref: "#/components/schemas/Meta" }
        errors:
          type: array
          items: { $ref: "#/components/schemas/ApiErrorDetail" }

    # ---- enums ----
...
    LoginRequest:
      type: object
      required: [tenantSlug, email, password]
      properties:
        tenantSlug: { type: string }
        email: { type: string, format: email }
        password: { type: string }
    MfaVerifyRequest:
      type: object
      required: [mfaToken, code]
      properties:
        mfaToken: { type: string }
        code: { type: string }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 12 }
    SessionTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: "#/components/schemas/AuthenticatedUser" }
    AuthenticatedUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        facilityId: { type: [string, "null"], format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { type: string }
        mustChangePassword: { type: boolean }
    LoginResult:
      type: object
      properties:
        requiresMfa: { type: boolean }
        mfaToken: { type: [string, "null"] }
        accessToken: { type: [string, "null"] }
        refreshToken: { type: [string, "null"] }
        user: { oneOf: [ { $ref: "#/components/schemas/AuthenticatedUser" }, { type: "null" } ] }
    EnvelopeLoginResult:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/LoginResult" }
        meta: { $ref: "#/components/schemas/Meta" }
    EnvelopeSessionTokens:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/SessionTokens" }
        meta: { $ref: "#/components/schemas/Meta" }
    EnvelopeAuthenticatedUser:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/AuthenticatedUser" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- claims ----
...
    OAuthTokenRequest:
      type: object
      required: [grant_type, client_id, client_secret]
      properties:
        grant_type: { type: string, const: client_credentials }
        client_id: { type: string }
        client_secret: { type: string }
        scope: { type: string, description: "Optional space-delimited subset (down-scoping)" }
    OAuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer }
        scope: { type: string }
      additionalProperties: false